requirements.txt
streamlit
pillow
scikit-learn
numpy
matplotlib
colorsys